♻️ refactor(inline-style-gallery): replace modal with inline gallery in Art Direction tab
- Embed style gallery/preview directly in Art Direction tab content area
- 3 display states: GALLERY (default when no art-direction), PREVIEW, RESULT
- Remove StyleGalleryModal component and all modal plumbing
- Remove list_styles auto-trigger (Ethical Design: no implicit UI changes)
- Apply upstream CORS fix: callback ref for ResizeObserver, dynamic slideCount

SPEC: 20260406-2226_inline-style-gallery
Progress: Tasks 1-7 complete, Task 8 (a11y verification) pending
