+++ web-ui/src/components/deck/SlideCarousel.tsx
@@ -37,11 +37,13 @@ interface SlideCarouselProps {
   specs?: SpecFiles | null
   /** Workflow phase detected from tool calls — drives spec tab auto-switch. */
   workflowPhase?: string | null
-  /** Callback to open the style gallery for re-selection. */
-  onStyleGalleryRequest?: () => void
+  /** Callback when user selects a style inline. */
+  onStyleSelect?: (name: string) => void
+  /** Cognito ID token for style API calls. */
+  idToken?: string
 }
 
-export function SlideCarousel({ slides, deckId, deckName, pptxUrl, isLoading, onSlideClick, scrollToSlide, onScrollComplete, headerActions, ownerAlias, specs, workflowPhase, onStyleGalleryRequest }: SlideCarouselProps) {
+export function SlideCarousel({ slides, deckId, deckName, pptxUrl, isLoading, onSlideClick, scrollToSlide, onScrollComplete, headerActions, ownerAlias, specs, workflowPhase, onStyleSelect, idToken }: SlideCarouselProps) {
   const slidesWithPreview = slides.filter((s) => s.previewUrl)
   const auth = useAuth()
   const [jsonLoading, setJsonLoading] = useState(false)
@@ -335,7 +337,8 @@ export function SlideCarousel({ slides, deckId, deckName, pptxUrl, isLoading, on
           content={specs?.[specTab] ?? null}
           specName={specTab.charAt(0).toUpperCase() + specTab.slice(1)}
           specKey={specTab}
-          onStyleChange={specTab === "artDirection" ? onStyleGalleryRequest : undefined}
+          onStyleSelect={specTab === "artDirection" ? onStyleSelect : undefined}
+          idToken={specTab === "artDirection" ? idToken : undefined}
         />
       )}
     </div>
